prov/psm2: Support building with latest PSM2 source code
The latest PSM2 source code introduces a Hardware Abstraction Layer
and as the result brings in changes to source code layout as well as
compiler flags. Amend the provider configuration code to detect the
new version of PSM2 source code to allow the provider to continue
be built with the PSM2 source included.

There is no impact if the provider is built separately from the PSM2
code, i.e., using PSM2 as a library.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>