NETCASSINI-920 - fix pid_idx handling
Parametrize calculations involving pid and pid_idx so they are
consistently derived from the number of bits in the VNI index (17) and
the number of bits used by the libfabric provider for the pid (9).

Correctly compute the pid_idx of the MR slots.

Check for out-of-range keys specified for MRs.

Signed-off-by: Joe Nemeth <jnemeth@cray.com>