fabric: Add fid_nic field directly to fi_info structure
It turns out that there are issues trying to assign fid_nic to
the existing handle field.  The handle field is ignored by
fi_dupinfo and fi_freeinfo, and as a result may point to
freed memory.  E.g. handle may reference a FI_CLASS_CONNREQ that
the provider may free on calling to fi_accept() or fi_reject().
So, it may not be safe to dereference.

To avoid this potential issue, a new field is added to fi_info.
This will allow fi_dupinfo/fi_freeinfo to handle processing it.
No application changes are needed for calls like fi_tostr() to
display the new NIC attributes, but the application must
recompile to get it.

This change results in the ABI needing to be updated.  A new
compatibility set of interfaces is the result.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>