prov/verbs: Fix completion handling
With FI_SELECTIVE_COMPLETION setting, the applications running with the
verbs provider still gets completion entries for RDMA and MSG messages without
a FI_COMPLETION flag.

The problem is fixed by setting VERBS_NO_COMP_FLAG to wr_id in case of
completion shouldn't be generated for this operation. When polling CQ,
this completions will be silently discarded.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>