prov/util: Add functions to increment counters.
Using a function pointer for each one of an EP's counters allows for the
removal of an if statement in the critical (and successful) path. By default,
the pointer is set to a noop.  The pointer is set to a function which
increments the counter only when the counter is bound to the EP.

Signed-off-by: yohann <yohann.burette@intel.com>