util/buf: Add tracking of buffers by index
Currently, there is no way to retrieve an already allocated buffer
by its index.
The patch implements way to retrieve a buffer by specifying its index.
util_buf_ftr structure is updated to contain index of the buffer and
util_buf_pool structure is updated to contain the regions table that is
dynamically expanded (if there is no space in the regions table, the
regions table is reallocated and `its size = its previous size + 16`).

The tracking by index functionality will be used for implementing
FI_AV_TABLE for the new version of the utility AV.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>