md2nroff: remove some useless code
The extra "git status"/"git add" is not necessary because a few lines
below, we just "git commit -a".  Also, with the new "all files in the
array contain the entire relative path name" schehem, we're not
currently chdir'ed into a git repository, so these git commands are
failing, anyway.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>