core/util_buf: Move slist entry to footer
This allows users to pre-format the data buffers and have
the information saved across free/alloc calls.  As part of
this change introduce some new helper functions to retrieve
the buffer footer given a pointer to the pool and user's
data buffer.  This results in the removal of union util_buf.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>