contrib/cray: compile fabtests with each build
If fabtests does not have the correct compile time target, it will
test against older build instead of the target build. This is not an easy
problem to resolve because ssh may not allow the propagation of environment
variables depending on system configuration. Rather than depend on
LD_LIBRARY_PATH, we will compile fabtests with each libfabric install instead.

Signed-off-by: James Swaro <jswaro@cray.com>