prov/rxm: Use MSG provider's inject write when iov_count=1
The main reasong of this commit is to fix a bug for `fi_writemsg(...,
FI_INJECT | ...)` with the message with iov_count or/and rma_iov_count > 1
and total size <= MSG provider's `inject_size`. RxM uses `fi_inject_write`
and `fi_inejct_writedata` for the iovec with index = 0 in this case.

This commit added check for iov_count and rma_iov_count before using MSG
inejct path. Fallback to fi_writemsg path if total_size > MSG `inject_size`
or FI_COMPLETION set or (iov_count != 1) or (rma_iov_count != 1)

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>