fabtests/hpcs: Add ring pattern using triggered ops.
This adds a new pattern that uses triggered ops in a ring-type
communication pattern.  A leader sends to the next node, which
triggers a send to the next node, and so on until the final
node sends a message back to the leader.

Multiple rings may execute concurrently.  (The maximum is to have
as many concurrent rings as ranks, with each node as a leader
of a single ring.)

The onesided and sendrecv tests have been updated to support
triggered ops, using either the old official API or the new
experimental deferred work API.  The new API is enabled with the
"--workqueue" test argument.

This changes some of the internal APIs.  Patterns now return
-ENODATA when they're done, so we can return other errno
codes in case there's an error.

Patterns also have an output "threshold" parameter, to indicate
to the caller how many receives should happen before a transmit is
triggered.

The test API now takes a "flags" parameter, to be used with
sendmsg, writemsg, etc...  When tests are run with triggered-ops
enabled patterns, the FI_TRIGGER flag will be set.

Signed-off-by: Jim Snow <jim.m.snow@intel.com>