prov/rxd: add FI_AV_MAP support and clarify internal addressing
Previously, FI_AV_TABLE was requested and the dg_addr was used to access the peer connection info via table. This approach was not compatible with AV_MAP. To allow for AV_MAP support, include an internal rxd_addr that maps to the dg_addr. Also includes an fi_addr table to map inserted addresses to the internal rxd_addr, while maintaining the FI_AV_TABLE for the app.

- Includes various renaming of variables to make the distinction between fi_addr, rxd_addr, and dg_addr clearer

Signed-off-by: aingerson <alexia.ingerson@intel.com>