prov/tcp: Simplify migration of passive socket to active
When opening a new endpoint, if a passive endpoint is passed
into the call, the attributes of the passive endpoint are
assumed by th active endpoint.

Rather than opening a new socket and copying the address from
the passive endpoint, we can simply shift the underlying socket
over to the passive endpoint.

To improve code clarity, adjust the check for whether the handle
references a passive endpoint by checking prior to casting the
handle to a more generic structure.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>