prov/verbs: Discard destination address when allocated PEP
RxM will allocate a passive endpoint using the fi_info structure
allocated for an active endpoint.  As a result, the info::
dest_addr field may be set.  This triggers the assertion in
fi_ibv_eq_cm_getinfo() that dest_addr is null.

The issue is that when the passive endpoint is created, the
fi_info passed in is copied to the PEP.  This includes the
dest_addr, if present.

It turns out that the xrc code discards the dest_addr.  Remove
the xrc portion of the check and always discard (i.e. ignore)
the dest_addr if present.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>