fabtests: Add test option to skip message allocations
The unexpected_msg test allocates its own data buffers
for transfers.  Add a test option, so that the common
code skips this step.

This fixes an issue with fi_close() failing because of
leaked MRs.

Apply this flag to the multi_recv test and replace the
indirect check for FI_MULTI_RECV support.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>