util/eq: Avoid possible NULL pointer dereference in eq_read
Problem reported by coverity.  Buf may be null, but is dereferenced
as part of tracking err_data.

Always free any saved error data when trying to read or peek
for a new error entry.  free() handles NULL, so there's no need
for a separate check.

Write the error data into the user's buffer prior to saving
the data.  This prevents trying to save data that is not
needed based on the application's requested version.  If the
err_data is copied into the user's buffer, the err_data
will be freed prior to trying to save it.  Also, do not save
err_data if the application is only peeking for the data.
This prevents freeing it prior to the application reading
it on a subsequent call.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>