prov/rxm: Fix handling of addresses inserted several times
This patch fixes the issues where application inserts the same
addresses several times. util/av searches the addresses in the
hash. If utl/av finds some address in the hash, it returns
address's fi_addr to the user.
So, relying on this fact, this fix is to check handles_av array
and retrieve the CMAP handle if this is known `fi_addr`.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>