prov/rxm: Remove rxm_ep_progress_one
This is a near duplicate call of rxm_ep_progress_multi.  The
latter is used throughout the code.  The progress one is only
called from the cq progress function, and only if the number
of completions per progress is 1.

This simplifies the code prior to fixing the EP synchronization
issues.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>