prov/rxm: Remove cmap lock processing directed receives
The EP lock serializes access to the posted receive queue.  The
cmap lock is unnecessary, or whatever it is attempting to protect
isn't clear from the code.

This avoids nesting the cmap and EP lock.  A follow on patch will
extend the use of the EP lock to ensure that message handling is
fully serialized.

As part of the lock cleanup, remove an incorrect comment.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>