pingpong: Remove OOB sync during connection setup
Using an OOB sync in the middle of the connection process
assumes that all connection messages are transferred after
calling connect and breaks manual progress mode.

The use of an OOB sync message is unnecessary in most of
the code paths.  Remove it everywhere except immediately
prior to starting the pingpong test, to better align the
performance timing.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>