pingpong: Fix manual progress support
When posting transmits, if we get EAGAIN, we must read
the CQ in order to drive progress.  However, the check
pp_get_cq_comp() skips this if the number of transfers
(0 in this case) equal the number of completions (also 0).

This change allows us to support manual progress and
support the rxm provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>