prov/rxd: remove restriction for potential double ack
There was an added restriction to prevent the sending of multiple acks for the same lost packet. However, in the specific case of an rma read bandwidth test where the ACK for the last send read request is dropped, this will result in a hang during finalize_ep because the receive side will never receive the ACK and will not be able to complete the received operation.
Remove this restriction so any dropped ACKs can be retransmitted if needed.

Signed-off-by: aingerson <alexia.ingerson@intel.com>