Fix sockets memory leak: free all posted rx
If the sockets ep is closed while there are still pending recvs, and
there are more recvs outstanding than the initial pool allows for,
then all the recvs malloc'd beyond the pool are leaked. There are
multiple possible solutions (including ignoring the problem because
there should never be pending recvs on close). I chose to simply
free them to avoid the leak.

Signed-off-by: Chris Dolan <chrisdolan@google.com>