RxM: fix null deref on aborted Rx.
If we exit abruptly, the underlying msg provider may return error
completions (e.g. -FI_ECANCELED) for all of the fi_recv calls that RxM
submits eagerly. Arguably, these should be dropped instead of
propagated to the client because they are an implementation detail of
the RxM provider, but in any case it's better not to try to
dereference the non-existent rx_buf->recv_entry pointer.

Signed-off-by: Chris Dolan <chrisdolan@google.com>