RxM: do not exceed rx limit for msg provider.
rx_attr->size is the size of the recv pool in the sockets provider. If we
prepost that many, then the pool is exhausted. When recvs complete, RxM
always preposts a new one before releasing the old one, so we're always
using one more buffer than the pool can support. There had been a leak in
the non-pool code for sockets that was triggered by this. A mitigation is
just to avoid exhausting the pool by pre-posting one less recv.

An alternative solution would be for RxM to cancel all preposted recvs on
close instead of just letting the core provider handle it.

Even with the sockets memory leak fixed, this is still better because we
won't ask for more than the core provider has declared it can support.

Signed-off-by: Chris Dolan <chrisdolan@google.com>