RxM: add support for FI_MR_RAW
This is a backward-compatible change to the RxM wire format for RMA
use cases which use large auth keys. If the core provider says that it
supports FI_MR_RAW _or_ the declared key size is larger than 64 bits
(when would you ever have the latter without the former??) then the
memory owner uses fi_mr_raw_attr to add the full RMA auth key and
base_addr to a new area in the packet header. The peer receives that
auth key and registers it temporarily via fi_mr_map_raw.

I chose to resend the key with every header, which is consistent with
the current implementation. One could imagine saving a some wire bytes
by sending it just once and remembering which peers you sent it to,
but that added complexity is unlikely to be worth a few bytes
especially if we'd need to support re-keying.

Caveat: this implementation assumes that the raw key size is a
constant and that all peers agree on it. This could be generalized by
including the key size in the header too, but that makes it trickier
to compute the available payload size in the packet because of the
variable-sized header.

Signed-off-by: Chris Dolan <chrisdolan@google.com>