pp: Remove FI_INJECT usage at the end of the test
This patch fixes the pingpong smoke test for the providers that has
`fi_ep_attr::inject_size` < 4 bytes. Now the pingpong doesn't use
FI_INJECT at the end of the test.

Also this patch removes excessive `memset()`s and defines the constant
for the latest buffer (+ verifies a RX buffer)

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>