fabtests/ubertest: add non-RX_CQ_DATA support
ft_x_ctrl.credits currently represent both the number of posted send/receives and also the number of completions. If a provider does not use the FI_RX_CQ_DATA mode bit, a received writedata operation will generate a completion but not consume a receive buffer. This causes the assert(credits <= max_credits) to fail since the credits will be incremented on the receive but never decremented when a receive buffer is posted successfully.

Fix this by using "credits" to only refer to the number of posted received buffers and track the number of received completions through the return of ft_comp_x. When a completion is found, adjust the credits according to the mode and flags of the completion entry.

Signed-off-by: aingerson <alexia.ingerson@intel.com>