Add FI_MR_RAW support to sockets provider.
To enable this, set an option like "FI_SOCKETS_MR_KEY_SIZE=16" to
override the default of 8 (= uint64_t). Values smaller than 8 are
ignored. All peers need to agree on this value implicitly or the TCP
stream will be appear corrupted.

On the memory-owner side, the key is composed of a 64-bit map key (as
before) plus rand() bytes to fill the remainder. The RMA validation
function now requires this key data when the key size is larger than
the default. When a peer invokes an RMA or atomic op, the peer adds
this raw key after the header but before any data payload.

The exhaustive domain_attr.mr_mode tests did not consider this
setting, so I masked off the FI_MR_RAW bit in the tests.

Signed-off-by: Chris Dolan <chrisdolan@google.com>