man/endpoint: Clarify counter and cq completions with FI_SELECTIVE_COMPLETION
This is an attempt to clarify and remove confusion around
the behavior of FI_SELECTIVE_COMPLETION, FI_INJECT, inject
operations, and counters.

The API and intended behavior are not changed.

The FI_SELECTIVE_COMPLETION examples are removed, as they apparently
just confuse people.  The man pages try to strict to precising defining
only the functions and what flags are valid for each call.

The wording around fi_ep_bind is moved around, so that all general
comments come first, followed by valid flags for ep->cq binding, then
flags for ep->cntr binding.  The more precise term 'cq completion' is
used in place of just 'completion' entry.

The counter man page is updated to indicate when a counter is
updated.

The inject call description is also updated to state the behavior
only, without referencing flags, as that results in confusion as well.
Other inject operations are updated to refer to fi_inject, in order
to avoid duplicating text.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>