prov/mrail: Fix infinite loop when handling ooo recv
When allocating an item for the ooo recv queue, the "next" field was
not reset to NULL and thus could lead to a loop in the queue after the
insertion of the new item. Later attempt to find the first match in the
queue would then enter an infinite loop. Fix by setting the "next" field
to NULL after allocation.

The issue only exists in non-debug build. With ENABLE_DEBUG defined
the "next" field is always set to NULL in the slist_remove_head()
function which is called for allocation.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>