prov/tcp: Merge process_srx_entry and process_rx_entry
This calls perform identical functionality, only differing by
where the posted receive occurs -- either to the EP or the SRX.
The difference only matters based on where the xfer_entry used
to track the posted receive comes from.  With srx, it entry
is allocated from the srx buffer pool.  For the EP, the entry
comes from the CQ's buffer pool.

We can use the new tcpx_rx_msg_release() call to return the
buffer to the correct pool, which allows us to merge the two
functions and avoid the duplicate code.

Include a fix to return the rx_entry to the correct buffer
pool as part of this merger.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>