util/bufpool: Replace 'GETTER' macros with functions
A set of macros are used to define a set of functions that
differ based on whether the underlying buffer pool is using
indexed buffers or not.  Replace the macros with functions.

This will be used by subsequent patches that fold in functions
that are only called from the current macros.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>