util/bufpool: Simplify attributes
Add flags to the buffer pool attributes.  The flags
will replace the mmap, track_used, and indexing ordered
fields.  The indexing used field is removed, as it doesn't
impact the code at all, since all buffers are associated
with an index.

As part of this change, we need to inspect calls that
create buffer pools.  Apply simple cleanups at those sites
to use the returned error code and fixup white spacing.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>