util/bufpool: Remove pool as input parameter where not needed
Now that he header can be retrieved using a direct offset, we
can remove the need to provide the pool as input into several
buffer pool calls.  This simplifies the API, and prevents bugs
that can result when the buffer is matched to the wrong pool.

Remove the pool parameter from all calls where it can be
obtained using the buffer header.  Update all calling sites.
For calling sites that result in a single line function, remove
the wrapper functions and call the buffer pool interfaces
directly.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>