runfabtests: Use exclude files by default if available
If an exclude file was not given, do a simple search for one
given the provider to use.  If no exclude file can be found,
run the test as normal.

As a result of this change, either an exclude file can be
specified, or an exclude list, but not both.

The main benefit is that by default the exclude files for the
tcp and udp providers will be used.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>