runfabtests: Use regular expressions for exclusion filter always
All of the existing exclude files require the use of regular expression
patterns.  Disable the -R option and just make this the default.

Note that the -R option is maintained as input, but just ignored.  This
avoids breaking external CI that passes -R as input.  Once external
CI is updated, we can remove the option completely.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>