prov/mrail: Fix how pass-through mode bits are checked
Modes are restrictions from the provider. The application can set more
mode bits than necessary in the hints to indicate that the application
is ready to handle those restrictions. It is incorrect to reject mode
bits outside the pass-through set because they may be cleared by the
core provider in the end. The correct way is to clear these bits from
hints so that the core provider would not assume those modes are handled
by upper layer.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>