prov/mrail: Allow returning multiple mrail info instances from fi_getinfo
When fi_getinfo is called into the mrail provider, the provider may discover
multiple choices for each rail. Previously a simple approach was taken to only
gather the first item from each rail to form a list for a single mrail info. A
special info node was added in front of the list as the "header". This approach
does provide the opportunity for applications to look into the details of each
rail (if the application is aware of the layout). However, it also has some
limitations:

(1) It may prevent some valid multi-rail provider layering from being presented
to the user, unless explicit provider name is given to fi_getinfo().

(2) It doesn't work properly with fi_dupinfo().

This patch mainly changes two things. First, for each mrail info, only the header
is returned to the user. The per-rail info details are kept private to the provider.
Second, multiple mrail info instances can be returned when available. This is made
possible by the first change.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>