prov/verbs: use credits to avoid CQ overrun
The Infiniband spec doesn't say when the space occupied by a send WR in send
queue space should be freed: after a WR is processed or when the corresponding
completion entry is read. If a particular hardware implements the former there
is a chance of overruning the CQ (send WRs processed at a faster rate than
completions read from CQ). So, based on the implementation, we may need to track
available space in CQ to avoid overrunning it (which renders any QPs bound to
it unusable).

To fix the issue, use an atomic counter to track available space and return
ENOMEM on ibv_post_send when there are no enough credits. Use function pointers
so that apps that run on hardware which won't have potential CQ overruns issue
aren't penalized.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>
(cherry picked from commit eb933bb80969c9c5dbebc07c5d71a51d5398b9a4)