prov/mrail: Ensure addresses inserted to address vector be distinguishable
The address vector is used as a convenient place to store "peer_info" and the
content of the "peer_info" is used as the "address" to insert into the address
vector. However, "peer_info" is initialized as all 0 and multiple insertion
would lead to the same entry in the address vector.

Fix by setting the "addr" field of "peer_info" before inserting to the address
vector.

Also fix the return value calculation when the "fi_addr" parameter is NULL.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>