prov/mrail: Bug fix - Missing completion for segmented receive
Add the FI_COMPLETION flag when claiming buffered messages from the lower
layer to ensure completion generation. In general we can't assume the
lower layer will generate the completion automatically without setting
the FI_COMPLETION flag.

For example, when rxm is used as the lower layer, completions are always
generated when claiming non-segmented messages because it is handled in
a routine specific to buffered receives. This somehow hides the issue.
However, if the message is segmented, the final step that finishes the
receive won't generate a completion if the FI_COMPLETION flag is not set.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>