prov/rxm: intialize variable that limits outstanding rndv tx requests
This patch is v1.7.x version of the fix that's added to master branch to avoid
updating headers used in CM data on a stable branch. A fixed value is used for
peer recevie queue. If the app uses default settings, the rx queue size shouldn't
be different for local and peer MSG endpoint.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>