prov/rxm: Ensure completion generation when claiming buffered messages
For non-segmented messages, the code path for claiming buffered receives
is different from regular receives and completions are always generated
there.  However, for segmented messages, the code to finishes the receive
is the same for buffered receives and regular receives and completion
generation depends on the FI_COMPLETION flag.

Add additional check to the finishing code so that completions are always
generated when buffered messages are claimed.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>