prov/verbs: Remove memory monitor
The verbs specific memory monitor will be replaced by one
managed by the core.

The verbs monitor is no longer operational because of the
prior changes to remove the malloc/free hooks, so this is
essentially removing unreachable code.

As part of this change, introduce a placeholder for the
userfaultfd memory monitor.  This is needed to limit our
changes without breaking the build.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>