core/mem_monitor: Change mem_monitor to callback model
The memory monitor code currently uses a polling model
(read from an event queue) to report changes to the
address space.  This is based on each memory region
being associated with a subscription.  Overlapping
regions would have separate subscriptions.

Replace this model with a callback mechanism.  Instead
of associating a region with a subscription, have the
caches register address ranges.  If an event occurs, the
monitor will callback each cache with the range that was
modified.  The caches are responsible for locating the
various regions that might have been affected.

This better aligns with reporting changes that occur to
a specific address range, independent of how many regions
that address range may belong to.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>