prov/rxm: refactor and fix cmap handle cleanup
Free cmap handles after CM thread join. Any unread events in the MSG provider
EQ should be cleaned up by MSG provider on endpoint close.

Also, refactor deleting connection handle. The actual deletion and reset is
postponed to RXM_CMAP_FREE notification from the event queue. Otherwise we
may miss to clean it up in rxm_cmap_free as the entry in peer list or handles_av
would be reset and we stop reading the EQ due to above change.

With this change we shouldn't hit failure in the assertion in
rxm_cmap_process_connect: assert(handle->state == RXM_CMAP_CONNREQ_RECV) since
we avoid reading the event queue on closing. The original issue is that the
CM thread didn't get a chance to process FI_CONNECTED event on a recv only
MSG endpoint and it's processed only at the very end when handle is deleted.
At this stage the handle state would be RXM_CMAP_SHUTDOWN and the assertion
fails.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>