prov/mail: CQ polling optimization for small size messages
CQ polling overhead increases in proportion to the number of rails in
use. For small messages, better performance can be achieved by staying
with a single rail.

This patch also adds an experimental runtime parameter to set the
configuration of rail utilization policies.

Some optimization measures are also applied to the CQ polling loop:

- Allow batch polling a rail when multiple completions are available

- When fixed rail is used for small messages, don't continue with other
  rails when at least one completion has been polled. This help improve
  small message performance.

- Always start with the rail that was polled successfully last time. This
  makes any rail to perform similarly as the default rail.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>