prov/rxd: remove unnecessary RXD_CANCELLED flag
If we find the target receive to cancel in the rx_list, it has not been matched to an incoming message, so there's no need to track whether it has been canceled or not. We can just complete the FI_CANCELLED error and free it.

If it is not in the list, it has either completed or is in progress of completing, so we treat it as a completed message

(cherry-picked from commit 9f6083030f964045b1e135bc6da0f371056761f5)

Signed-off-by: aingerson <alexia.ingerson@intel.com>