prov/rxd: fix packet type declaration and printing
Because RxD does not use two of the ofi_op enums in ofi_proto.h and then defines RxD-specific ones after, the printing (which is based on the array index of the new defines) is off by 2 for the new RxD-specific enums.

To fix this, eliminate the connection to the ofi_proto.h enums and define completely separate RXD types but make sure to align them with the ofi_proto enums so we don't break the protocol. Add in 2 placeholder enums to line up the enums for printing

(cherry-picked from commit ee2e622a9bc950d43e3d37b0e70ca037262a715c)

Signed-off-by: aingerson <alexia.ingerson@intel.com>